# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gacha Collection Project - The Dubliners",
    "description": "Specifiche API del progetto divise per microservizi.",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://localhost"
    },
    {
      "url": "https://localhost:444"
    }
  ],
...
  "paths": {
...
    "/auction_service/auctions/active": {
      "get": {
        "tags": [
          "Gacha Auction Service"
        ],
        "summary": "Get active auctions",
        "description": "Retrieves a list of all active auctions currently available.",
        "responses": {
          "200": {
            "description": "Active auctions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "auction_id": {
                        "type": "integer",
                        "description": "The unique ID of the auction.",
                        "example": 1
                      },
                      "gacha_id": {
                        "type": "integer",
                        "description": "The unique ID of the gacha being auctioned.",
                        "example": 101
                      },
                      "issuer_id": {
                        "type": "integer",
                        "description": "The ID of the user who issued the auction.",
                        "example": 55
                      },
                      "current_user_winner_id": {
                        "type": "integer",
                        "description": "The ID of the user currently winning the auction.",
                        "example": 3
                      },
                      "current_bid": {
                        "type": "number",
                        "format": "float",
                        "description": "The current highest bid for the auction.",
                        "example": 150
                      },
                      "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "description": "The timestamp indicating when the auction started.",
                        "example": "2024-11-17T12:00:00Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No active auctions found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "No active auctions found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An internal error occurred"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auction_service/players/{userId}/setAuction": {
      "post": {
        "tags": [
          "Gacha Auction Service"
        ],
        "summary": "Create a new auction",
        "description": "Allows a user to auction one of their gacha items. The auction will last for a specified time (default is 1 minute).",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the user setting the auction."
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The JWT token used for authentication."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gacha_id": {
                    "type": "integer",
                    "description": "The unique ID of the gacha being auctioned.",
                    "example": 101
                  },
                  "base_price": {
                    "type": "number",
                    "format": "float",
                    "description": "The starting price of the auction.",
                    "example": 50
                  }
                },
                "required": [
                  "gacha_id",
                  "base_price"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Auction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Auction created successfully"
                    },
                    "auction_id": {
                      "type": "integer",
                      "description": "The unique ID of the newly created auction.",
                      "example": 2001
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid input data"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access. Token is missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token is missing or invalid"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Gacha not found or not owned by user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Gacha not found or not owned by user"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error or communication error with gacha service",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": {
                        "serviceError": {
                          "value": "An error occurred while communicating with the gacha service"
                        },
                        "internalError": {
                          "value": "An internal error occurred"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Gacha service is temporarily unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Gacha service is temporarily unavailable"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auction_service/auctions/{auctionID}/bid": {
      "post": {
        "tags": [
          "Gacha Auction Service"
        ],
        "summary": "Place a bid on an auction",
        "description": "Allows a user to place a bid on a specific auction.",
        "parameters": [
          {
            "name": "auctionID",
            "in": "path",
            "description": "The ID of the auction to place a bid on",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The JWT token used for authentication."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "description": "The ID of the user placing the bid"
                  },
                  "bid_amount": {
                    "type": "number",
                    "format": "float",
                    "description": "The bid amount (must be higher than the current bid)"
                  }
                },
                "required": [
                  "user_id",
                  "bid_amount"
                ],
                "example": {
                  "bid_amount": 500,
                  "user_id": 1
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bid placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bid placed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, auction is not active, insufficient funds, or bid amount is too low",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": {
                        "invalidAuctionId": {
                          "value": "AuctionId must be a positive number"
                        },
                        "invaludInput": {
                          "value": "Invalid input data"
                        },
                        "amountError": {
                          "value": "Amount must be a positive number"
                        },
                        "invalidUser": {
                          "value": "User_id must be a positive number"
                        },
                        "bidError": {
                          "value": "Bid amount must be higher than the current bid"
                        },
                        "insufficientFounds": {
                          "value": "insufficientFounds"
                        },
                        "auctionActive": {
                          "value": "Auction is no longer active"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access. Token is missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token is missing or invalid"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Auction not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Auction not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error or issues communicating with external services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An error occurred while communicating with other services"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Authentication service is temporarily unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Authentication service is temporarily unavailable"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}